+++ app/src/background.js
@@ -1,3 +1,5 @@
+//For intersect we are not switching to debounce
+//as the callback will handle multiple elements entering the sigularly
 const INTERSECTION_UPDATE_STEP_VIDEOS = 2;
 let allIntersectVideoElements = null;
 const intersectionObserverOtherVideos = new IntersectionObserver(
@@ -9,6 +11,8 @@ const intersectionObserverOtherVideos = new IntersectionObserver(
   },
 );
 
+//For intersect we are not switching to debounce
+//as the callback will handle multiple elements entering the sigularly
 const INTERSECTION_UPDATE_STEP_SHORTS = 2;
 let allIntersectShortElements = null;
 const intersectionObserverOtherShorts = new IntersectionObserver(
@@ -211,6 +215,24 @@ async function untranslateCurrentMobileVideoDescriptionHeader() {
     fakeNodeID,
     originalNodeSelector,
     () => document.location.href,
+    "span",
+    true,
+  );
+}
+
+// Untranslate channel featured video title for the mobile (m.youtube.com) layout
+async function untranslateCurrentMobileFeaturedVideoChannel() {
+  if (!window.YoutubeAntiTranslate.isMobile()) {
+    return;
+  }
+  const fakeNodeID =
+    "yt-anti-translate-fake-node-mobile-featured-video-channel";
+  const originalNodeSelector = `ytm-channel-featured-video-renderer > a > h3 > span.yt-core-attributed-string:not(#${fakeNodeID})`;
+
+  await createOrUpdateUntranslatedFakeNode(
+    fakeNodeID,
+    originalNodeSelector,
+    (el) => el.closest("a").href,
     "span",
     true,
   );
@@ -646,6 +668,8 @@ async function untranslate() {
   const otherShortsPromise = untranslateOtherShortsVideos(); // Call the new function
   const currentMobileVideoDescriptionPromise =
     untranslateCurrentMobileVideoDescriptionHeader();
+  const currentMobileFeaturedVideoChannel =
+    untranslateCurrentMobileFeaturedVideoChannel();
 
   // Wait for all promises to resolve concurrently
   await Promise.all([
@@ -658,6 +682,7 @@ async function untranslate() {
     currentShortVideoLinksPromise,
     otherShortsPromise,
     currentMobileVideoDescriptionPromise,
+    currentMobileFeaturedVideoChannel,
   ]);
 
   // update intersect observers