+++ tests/helpers/SetupUBlockAndAuth.ts
@@ -92,7 +92,7 @@ export async function setupUBlockAndAuth(
         });
 
         if (!localeLoaded) {
-          await page.goto("https://www.youtube.com/@MrBeast");
+          await page.goto("https://www.youtube.com/feed/you");
 
           try {
             await page.waitForLoadState("networkidle", { timeout: 5000 });